refactor: remove omnichainGovernance feature flag
