+++ .changeset/grumpy-lies-sparkle.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": patch
+---
+
+remove omnichainGovernance feature flag
