+++ apps/evm/src/hooks/useIsFeatureEnabled/index.tsx
@@ -117,22 +117,6 @@ export const featureFlags = {
     ChainId.BASE_MAINNET,
     ChainId.BASE_SEPOLIA,
   ],
-  omnichainGovernance: [
-    ChainId.BSC_TESTNET,
-    ChainId.BSC_MAINNET,
-    ChainId.SEPOLIA,
-    ChainId.ETHEREUM,
-    ChainId.OPBNB_TESTNET,
-    ChainId.OPBNB_MAINNET,
-    ChainId.ARBITRUM_SEPOLIA,
-    ChainId.ARBITRUM_ONE,
-    ChainId.ZKSYNC_SEPOLIA,
-    ChainId.ZKSYNC_MAINNET,
-    ChainId.OPTIMISM_MAINNET,
-    ChainId.OPTIMISM_SEPOLIA,
-    ChainId.BASE_MAINNET,
-    ChainId.BASE_SEPOLIA,
-  ],
   gaslessTransactions: [ChainId.ZKSYNC_SEPOLIA, ChainId.ZKSYNC_MAINNET],
   web3DomainNames: [ChainId.BSC_MAINNET, ChainId.ETHEREUM, ChainId.ARBITRUM_ONE],
 };
